Issues with volume-mounting config files

+++ makefile
@@ -1,15 +1,26 @@
 # makefile
 
 raspap: .build
+	file=/etc/dnsmasq.conf
+
+	if [ ! -e "${file}" ] ; then
+	    touch "${file}"
+	fi
+
+	if [ ! -w "${file}" ] ; then
+	    echo cannot write to ${file}
+	    exit 1
+	fi
 ## volume-mount all the networking things
 	-@docker rm -f $@ > /dev/null
 	docker run -d \
+		--privileged \
 		--name $@ \
 		-p 80:80 \
 		--net=host \
-		-v /etc/network/interfaces \
-		-v /etc/hosts \
-		-v /etc/dnsmasq.conf \
+		-v /etc/network/interfaces:/etc/network/interfaces \
+		-v /etc/hosts:/etc/hosts \
+		-v /etc/dnsmasq.conf:/etc/dnsmasq.conf \
 		raspap
 
 build: .build
